+++ file-sharing/src/tabs/iSCSI/ui/screens/radosBlockDeviceManagement/RBDTableEntry.vue
@@ -2,7 +2,7 @@
   <tr>
     <td>{{ getDeviceName() }}</td>
     <td>{{ getDeviceClassName() }}</td>
-    <td>{{ _cockpit.format_bytes(device.maximumSize) }}</td>
+    <td>{{ formatBytes(device.maximumSize ?? 0, "binary") }}</td>
     <td class="button-group-row justify-end">
       <button @click="showExpandScreen = true">
         <span class="sr-only">Resize</span>
@@ -25,12 +25,11 @@ import { WrenchIcon } from "@heroicons/vue/20/solid";
 import { RadosBlockDevice } from "@/tabs/iSCSI/types/cluster/RadosBlockDevice";
 import { LogicalVolume } from "@/tabs/iSCSI/types/cluster/LogicalVolume";
 import { ref } from 'vue';
+import { formatBytes } from "@45drives/houston-common-lib";
 import RBDExpansionScreen from "@/tabs/iSCSI/ui/screens/radosBlockDeviceManagement/RBDExpansionScreen.vue";
 import { Modal } from "@45drives/houston-common-ui";
 import { PlusIcon } from "@heroicons/vue/24/solid";
 import { VirtualDevice } from '@/tabs/iSCSI/types/VirtualDevice';
-
-const _cockpit = cockpit;
 
 const props = defineProps<{
   device: RadosBlockDevice | LogicalVolume;