+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## Cockpit File Sharing 4.6.1-2
+## Cockpit File Sharing 4.6.2-2
 
-* Repackage with correct houston-common submodule commit
+* building test package